Refactor: Modularize codebase and improve code structure
